+++ src/pages/Notice/index.tsx
@@ -1,4 +1,4 @@
-import { getNoticeListData, getNoticeSearchData } from 'apis/ShareTogether';
+import { getNoticeListData, getNoticeSearchData } from 'apis';
 import NoticeHeader from 'components/organisms/NoticeHeader';
 import NoticeList from 'components/organisms/NoticeList';
 import SeachForm from 'components/organisms/SearchForm';
