+++ src/pages/Setting/index.tsx
@@ -1,12 +1,12 @@
 import { withStyles } from '@material-ui/core/styles';
 import Switch from '@material-ui/core/Switch';
+import { getUserInfo, updateUserInfo } from 'apis';
 import editIcon from 'assets/images/editwhite.png';
 import xIcon from 'assets/images/xIcon.svg';
 import StyledInput from 'components/atoms/StyledInput';
 import DropDown from 'components/molecules/DropDown';
 import DropDownMulti from 'components/molecules/DropDownMulti';
 import Header from 'components/organisms/Header';
-import { getUserInfo, updateUserInfo } from 'libs/axios';
 import React, { useEffect, useRef, useState } from 'react';
 import { useHistory, withRouter } from 'react-router-dom';
 import { useRecoilState, useRecoilValue } from 'recoil';
